1. Refactored AddPollDto, AddReviewDto to reduce redundancy, remove unused methods and improve code efficiency. 2. Removed AddSoftAskAnswerDto.java, AnswerCount.java 3. Refactored AddSoftAskDto and AddSoftAskReplyDto to set parent reply to a reply on a soft ask. 4. Updated AuthenticationServiceImpl and AuthenticationServiceTest. 5. Updated BaseLinkDto to set message key and removed unused methods. 6. Created CalendarMapper and CalendarMapperImpl. 7. Updated CalendarRepository findByStatus() and findMany() methods to search by CalendarStatus enum. 8. Refactored CalendarResponse to remove isActive field and replace with CalendarStatusInfo. 9. Refactored CalendarSearchRequest to return List of calendar status to filter and search for calendars. 10. Refactored CalendarServiceImpl and extract to create method to find calendar. 11. Renamed NotAnAdminOfChatSpaceException.java to ChatSpaceNotAnAdminException, and updated usage in ChatSpaceController. 12. Refactored ChatSpaceEventServiceImpl to improve reuse of shared and existing code. 13. Refactored ChatSpaceJoinServiceImpl. 14. Updated ChatSpaceMapper and ChatSpaceMapperImpl to set bookmark, like and total members info. 15. Refactored ChatSpaceMember and ChatSpaceMemberMapper and ChatSpaceMemberMapperImpl. 16. Refactored ChatSpaceMemberServiceImpl, ChatSpaceOperationsService, and ChatSpaceOperationsServiceImpl. 17. Updated ChatSpaceRepository to decrement and increment bookmark and like count. 18. Added getter in ChatSpaceRequestToJoinPendingSelect. 19. Refactored ChatSpaceRequestToJoinStatus. 20. Refactored and updated ChatSpaceResponse to contain bookmark, like and share info. 21. Updated ChatSpaceSearchController. 22. Updated ChatSpaceSearchService and ChatSpaceSearchServiceImpl and ChatSpaceService and ChatSpaceServiceImpl. 23. Refactored CommonStreamOtherServiceImpl and CommonStreamService. 24. Removed DeleteSoftAskAnswerDto.java and refactored DeleteSoftAskReplyDto. 25. Added util methods in HasId including getIdsContainIn(), getIdsNotContainIn() and getIds() to use on a collection. 26. Added updateType() in Like. 27. Added methods in Likeable and extends HasId. 28. Added documentation in LikeController. 29. Updated LikeParentType to contain Business and LikeResponse. 30. Updated LikeService and refactored LikeServiceImpl. 31. Refactored and updated Link to contain Business domain, also LinkController documentation updated. 32. Updated LinkMapperImpl. 32. Updated LinkRepository to find links by a business. 33. Updated LinkSearchRequest to add getters to retrieve parentId and parent type. 34. Replaced Object with Long type in LinkSearchResult. 35. Updated LinkService and LinkServiceImpl. 36. Updated LinkType to contain community/group format and business format. 37. Updated LinkTypeInfo to contain business format. 38. Refactored MaskedStreamLinkUri.of() to return null if value is null. 39. Updated i18n documentation for responses, info and errors. 40. Move static methods from MiscServiceImpl to HasId. 41. Added share count to PollResponse and Poll. 42. Updated spring boot version to 3.5.4 43. Updated ProcessRequestToJoinChatSpaceResponse. 44. Set properties on ReviewResponse in ReviewMapperImpl. 45. Removed all soft ask answer related models. 46. Added method getSoftAskType() in SoftAskCommonResponse. 47. Updated SoftAskContentUpdateResponse. 48. Refactored SoftAskInfoMapper and SoftAskInfoMapperImpl to remove unused methods. 49. Refactored SoftAskOperationService and SoftAskOperationServiceImpl. 50. Refactored SoftAskParentDetailHolder and SoftAskReplyAddResponse to contain child replies count. 51. Updated SoftAskReplyRepository to handle update of bookmark count. 52. Updated SoftAskReplyResponse to contain share count, reply count and replies and author infos. 53. Refactored SoftAskSearchRequest, SoftAskRepository. 54. Updated domain classes to add share count, vote count, reply count and child reply count in SoftAsk, SoftAskReply, FleenStream, ChatSpace and Review.

+++ src/main/java/com/fleencorp/feen/chat/space/model/response/member/LeaveChatSpaceResponse.java
@@ -30,14 +30,14 @@ public class LeaveChatSpaceResponse extends LocalizedResponse {
   private ChatSpaceMembershipInfo membershipInfo;
 
   @JsonProperty("total_members")
-  private Long totalMembers;
+  private Integer totalMembers;
 
   @Override
   public String getMessageCode() {
     return "leave.chat.space";
   }
 
-  public static LeaveChatSpaceResponse of(final Long chatSpaceId, final ChatSpaceMembershipInfo membershipInfo, final Long totalMembers) {
+  public static LeaveChatSpaceResponse of(final Long chatSpaceId, final ChatSpaceMembershipInfo membershipInfo, final Integer totalMembers) {
     return new LeaveChatSpaceResponse(chatSpaceId, membershipInfo, totalMembers);
   }
 }
